fix: move event slug helpers to utils.ts file to avoid API key error
The UI was sometimes displaying an error, presumably because when importing
events.ts it was importing base, which requires the API key and it was not
loaded. Probably this was happening only in client code, I did not try to debug
it in detail once I figured out which commit was the originator of the error.