fix: weight 'interested' as 4x 'maybe' for sorting
This depends a lot on the specifics and is somewhat subjective but thinking
about it some more it seems plausible to me that a person who voted
'interested' is four times as likely to attend as one who voted 'maybe'.
Sorting by total votes is supposed to give hosts a sense of the (relative)
popularity of their proposal.