Fix RSVPing from mobile
It was impossible to (un-)RSVP from mobile because clicking on a session block brought up the user select modal, but clicking the button did nothing. This commit just makes mobile behave the same as non-mobile; in upstream, clicking on a session block brings up the user select modal on mobile, too, though the modal's RSVP button was not bugged there.