Set the combobox as non-multiple when only a single user is meant to be selected.
This means that in the pages for an event's sessions and proposals, when selecting a user, after choosing one the combobox automatically closes. I am not aware of any other effects that it has.

The types could potentially be improved - when selectMany is false, hosts and setHosts can be only `Guest?` instead of `Guest[]`.